All Account interfaces now extend a new interface, Account.java. This contains common methods used through all inheriting Account interfaces. Thanks to Hugo5551 for suggesting this.
Currencies' starting balances can now be retrieved through specifying a nullable player UUID and nullable world name. They should be null in the case the starting balance is not being retrived for a Player account, otherwise it is expected they are specified.
Thanks to Hugo5551 for suggesting this.

Added a new event, GeneralAccountTransactionEvent. This is expected to be fired by all transactions from all Account types, in addition to the specific ones.
Thanks to Hugo5551 for suggesting this.

Finished the meat of all subcommands, they should be functional now, just without nice looking customisable messages which will come later on.

I've added a Migrate subcommand which allows users to migrate balances of accounts from one economy provider to another.
Thanks to Hugo5551 for suggesting this.

Took 1 hour 16 minutes