Fix an issue with PlayerAccount#setPermission
I didn't notice this when writing 2c47e92 , but the problem comes down to
" the consumer plugin waits for completion, doesn't receive one ".
That's why I implemented a fail reason and called the fail method.
I didn't think calling success is appropriate here ; we're not modifying
anything, soo.

Also added TriState#fromBoolean method because someone forgot to add it when
creating this enum ~~(definitely not me)~~