Highly optimise EventCaller
previously we had the tree set because i thought that it would be better to
get stuff auto sorted without too much of a hassle. well turns out the old
implementation was not _that_ fast. i've changed all sets to lists in
EventCaller. also that now we have a custom SortedList which sorts on each
add. another pro of lists is that now we don't need to do copies and stuff
we can just store an index and call it a day. just see the changes for
yourself :)

also added a push-to-the-limits test which registers 100 listeners into
the event bus, some of which blocking.