Replace EconomyTransactionInitiator with Cause
The EconomyTransactionInitiator interface was confusing and kinda hard to work with. The new Cause interface is abstracted out from the economy API to allow usages in other future APIs (the permissions API for example). It is also more OOP friendly compared to the EconomyTransactionInitiator, and much more straightforward to work with than the EconomyTransactionInitiator.